tests: extract duplicated functionality
JvmTest shared functionality and there was no need
